+++ tests/memory.c
@@ -13,13 +13,13 @@
 TEST_LOGGER3(meminfo)
 {
     RBRInstrumentMemoryInfo expected = {
-        .dataset = 1,
+        .dataset = RBRINSTRUMENT_DATASET_STANDARD,
         .used = 1528,
         .remaining = 134216192,
         .size = 134217728
     };
     RBRInstrumentMemoryInfo actual = {
-        .dataset = 1
+        .dataset = RBRINSTRUMENT_DATASET_STANDARD
     };
 
     TestIOBuffers_init(buffers,
@@ -35,6 +35,21 @@ TEST_LOGGER3(meminfo)
     TEST_ASSERT_EQ(expected.size, actual.size, "%" PRIi32);
     TEST_ASSERT_STR_EQ("meminfo dataset = 1" COMMAND_TERMINATOR,
                        buffers->writeBuffer);
+
+    return true;
+}
+
+TEST_LOGGER3(meminfo_invalid_dataset)
+{
+    RBRInstrumentMemoryInfo test = {
+        .dataset = RBRINSTRUMENT_UNKNOWN_DATASET
+    };
+
+    TestIOBuffers_init(buffers, "", 0);
+    RBRInstrumentError err = RBRInstrument_getMemoryInfo(instrument, &test);
+    TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_INVALID_PARAMETER_VALUE,
+                        err,
+                        RBRInstrumentError);
 
     return true;
 }
@@ -105,6 +120,24 @@ TEST_LOGGER3(readdata)
     TEST_ASSERT_STR_EQ("readdata dataset = 1, size = 1400, offset = 2800"
                        COMMAND_TERMINATOR,
                        buffers->writeBuffer);
+
+    return true;
+}
+
+TEST_LOGGER3(readdata_invalid_dataset)
+{
+    RBRInstrumentData test = {
+        .dataset = RBRINSTRUMENT_UNKNOWN_DATASET,
+        .size    = 0,
+        .offset  = 0,
+        .data    = NULL
+    };
+
+    TestIOBuffers_init(buffers, "", 0);
+    RBRInstrumentError err = RBRInstrument_readData(instrument, &test);
+    TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_INVALID_PARAMETER_VALUE,
+                        err,
+                        RBRInstrumentError);
 
     return true;
 }
@@ -233,6 +266,168 @@ TEST_LOGGER3(memformat_newtype_set)
         RBRINSTRUMENT_MEMFORMAT_CALBIN00);
     TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_SUCCESS, err, RBRInstrumentError);
     TEST_ASSERT_STR_EQ(command, buffers->writeBuffer);
+
+    return true;
+}
+
+TEST_LOGGER3(postprocessing)
+{
+    RBRInstrumentPostprocessing expected = {
+        .status = RBRINSTRUMENT_POSTPROCESSING_STATUS_IDLE,
+        .channels = {
+            .count = 5,
+            .channels = {
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_MEAN,
+                    .label = "pressure_01"
+                },
+                {
+                    .function =
+                        RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_SAMPLE_COUNT,
+                    .label = "pressure_01"
+                },
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_MEAN,
+                    .label = "temperature_01"
+                },
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_STD,
+                    .label = "temperature_01"
+                },
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_MEAN,
+                    .label = "conductivity_01"
+                }
+            }
+        },
+        .binReference = "pressure_01",
+        .binFilter = RBRINSTRUMENT_POSTPROCESSING_BINFILTER_NONE,
+        .binSize = 50.0,
+        .tstampMin = RBRINSTRUMENT_DATETIME_MIN,
+        .tstampMax = RBRINSTRUMENT_DATETIME_MAX,
+        .depthMin = 10.0,
+        .depthMax = 1000.0
+    };
+    RBRInstrumentPostprocessing actual;
+
+    const char *command = "postprocessing status = idle, channels = "
+                          "mean(pressure_01)|count(pressure_01)"
+                          "|mean(temperature_01)|std(temperature_01)"
+                          "|mean(conductivity_01), "
+                          "tstamp_min = 20000101000000, "
+                          "tstamp_max = 20991231235959, "
+                          "binsize = 50.0, binreference = pressure_01, "
+                          "depth_min = 10.0, depth_max = 1000.0, "
+                          "binfilter = none" COMMAND_TERMINATOR;
+    TestIOBuffers_init(buffers, command, 0);
+    RBRInstrumentError err = RBRInstrument_getPostprocessing(instrument,
+                                                             &actual);
+    TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_SUCCESS, err, RBRInstrumentError);
+    TEST_ASSERT_ENUM_EQ(expected.status,
+                        actual.status,
+                        RBRInstrumentPostprocessingStatus);
+    TEST_ASSERT_EQ(expected.channels.count, actual.channels.count, "%" PRIi32);
+    for (int i = 0; i < expected.channels.count; i++)
+    {
+        TEST_ASSERT_ENUM_EQ(expected.channels.channels[i].function,
+                            actual.channels.channels[i].function,
+                            RBRInstrumentPostprocessingAggregate);
+        TEST_ASSERT_STR_EQ(expected.channels.channels[i].label,
+                           actual.channels.channels[i].label);
+    }
+    TEST_ASSERT_STR_EQ(expected.binReference, actual.binReference);
+    TEST_ASSERT_ENUM_EQ(expected.binFilter,
+                        actual.binFilter,
+                        RBRInstrumentPostprocessingBinFilter);
+    TEST_ASSERT_EQ(expected.binSize, actual.binSize, "%f");
+    TEST_ASSERT_EQ(expected.tstampMin, actual.tstampMin, "%" PRIi64);
+    TEST_ASSERT_EQ(expected.tstampMax, actual.tstampMax, "%" PRIi64);
+    TEST_ASSERT_EQ(expected.depthMin, actual.depthMin, "%f");
+    TEST_ASSERT_EQ(expected.depthMax, actual.depthMax, "%f");
+    TEST_ASSERT_STR_EQ("postprocessing all" COMMAND_TERMINATOR,
+                       buffers->writeBuffer);
+
+    return true;
+}
+
+TEST_LOGGER3(postprocessing_set)
+{
+    RBRInstrumentPostprocessing postprocessing = {
+        .status = RBRINSTRUMENT_UNKNOWN_POSTPROCESSING_STATUS,
+        .channels = {
+            .count = 5,
+            .channels = {
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_MEAN,
+                    .label = "pressure_01"
+                },
+                {
+                    .function =
+                        RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_SAMPLE_COUNT,
+                    .label = "pressure_01"
+                },
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_MEAN,
+                    .label = "temperature_01"
+                },
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_STD,
+                    .label = "temperature_01"
+                },
+                {
+                    .function = RBRINSTRUMENT_POSTPROCESSING_AGGREGATE_MEAN,
+                    .label = "conductivity_01"
+                }
+            }
+        },
+        .binReference = "pressure_01",
+        .binFilter = RBRINSTRUMENT_POSTPROCESSING_BINFILTER_NONE,
+        .binSize = 50.0,
+        .tstampMin = RBRINSTRUMENT_DATETIME_MIN,
+        .tstampMax = RBRINSTRUMENT_DATETIME_MAX,
+        .depthMin = 10.0,
+        .depthMax = 1000.0
+    };
+
+    const char *command = "postprocessing binreference = pressure_01, "
+                          "binfilter = none, binsize = 50.0" COMMAND_TERMINATOR
+                          "postprocessing tstamp_min = 20000101000000"
+                          COMMAND_TERMINATOR
+                          "postprocessing tstamp_max = 20991231235959"
+                          COMMAND_TERMINATOR
+                          "postprocessing depth_min = 10.0, depth_max = 1000.0"
+                          COMMAND_TERMINATOR
+                          "postprocessing channels = mean(pressure_01)"
+                          "|count(pressure_01)|mean(temperature_01)"
+                          "|std(temperature_01)|mean(conductivity_01)"
+                          COMMAND_TERMINATOR;
+
+    TestIOBuffers_init(buffers, command, 0);
+    RBRInstrumentError err = RBRInstrument_setPostprocessing(instrument,
+                                                             &postprocessing);
+    TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_SUCCESS, err, RBRInstrumentError);
+    TEST_ASSERT_STR_EQ(command, buffers->writeBuffer);
+
+    return true;
+}
+
+TEST_LOGGER3(postprocessing_command)
+{
+    const char *command = "postprocessing status = processing"
+                          COMMAND_TERMINATOR;
+    TestIOBuffers_init(buffers, command, 0);
+
+    RBRInstrumentPostprocessingStatus result;
+    RBRInstrumentError err = RBRInstrument_setPostprocessingCommand(
+        instrument,
+        RBRINSTRUMENT_POSTPROCESSING_COMMAND_START,
+        &result);
+    TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_SUCCESS, err, RBRInstrumentError);
+    TEST_ASSERT_ENUM_EQ(RBRINSTRUMENT_POSTPROCESSING_STATUS_PROCESSING,
+                        result,
+                        RBRInstrumentPostprocessingStatus);
+    TEST_ASSERT_STR_EQ("postprocessing command = start" COMMAND_TERMINATOR,
+                       buffers->writeBuffer);
 
     return true;
 }